+++ src/main/java/gui/IconButton/searchButtonTop.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.IconButton;
 
 import search.Search;
